requirements.txt
streamlit
groq
fpdf
streamlit-ace
python-dotenv
